requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.1.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.1
cattrs==24.1.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudscraper==1.2.71
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cryptography==44.0.1
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.12
decorator==5.1.1
deepgram-sdk==2.12.0
deprecation==2.1.0
distro==1.9.0
docstring-to-markdown==0.15
docstring_parser==0.16
easyocr==1.7.2
elevenlabs==2.3.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
faiss-cpu==1.7.4
fastapi==0.115.11
filelock==3.17.0
fonttools==4.58.2
fpdf==1.7.2
frozenlist==1.5.0
fsspec==2025.2.0
genai==2.1.0
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-aiplatform==1.79.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-genai==1.0.0
google-resumable-media==2.7.2
googleapis-common-protos==1.67.0rc1
greenlet==3.1.1
grpc-google-iam-v1==0.14.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
imageio==2.37.0
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
jedi-language-server==0.43.1
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
langchain==0.3.18
langchain-community==0.3.17
langchain-core==0.3.41
langchain-openai==0.3.7
langchain-text-splitters==0.3.6
langsmith==0.3.4
lazy_loader==0.4
llvmlite==0.44.0
lsprotocol==2023.0.1
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.0
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.24.2
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.3
nltk==3.9.1
noisereduce==3.0.3
numba==0.61.0
numpy==1.26.4
openai==0.28.0
openai-whisper==20240930
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parso==0.8.4
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
platformdirs==4.3.6
playwright==1.50.0
pluggy==1.5.0
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
psutil==6.1.1
pure_eval==0.2.3
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
pyee==12.1.1
pygls==1.3.1
Pygments==2.19.1
PyMuPDF==1.25.3
pyparsing==3.2.3
pypdfium2==4.30.1
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.12.0
python-multipart==0.0.20
pytz==2025.1
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
shapely==2.0.7
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.6
SpeechRecognition==3.14.1
SQLAlchemy==2.0.37
stack-data==0.6.3
starlette==0.46.1
streamlit==1.41.1
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
tifffile==2025.2.18
tiktoken==0.9.0
toml==0.10.2
torch==1.13.1+cpu
torchvision==0.14.1+cpu
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
wcwidth==0.2.13
websockets==14.2
yarl==1.18.3
zstandard==0.23.0
